+++ NEWS.md
@@ -21,6 +21,8 @@ mutation type information to the names of files generated.
 
 * Added new exported data `signature.params`.
 
+* Added new exported data `ID.MMR.params`.
+
 ### Fixed
 * Fixed a bug in function `GenerateSynExposureOneSample` to round the mutations
 due to each signature.