+++ NEWS.md
@@ -16,6 +16,9 @@ have empirical signature parameters from all cancer types.
 * Updated function `GenerateSyntheticTumors` and `GenerateNoisyTumors` to append
 mutation type information to the names of files generated.
 
+* Updated function `GenerateSyntheticExposures` to remove signatures that have zero
+exposure in the synthetic data.
+
 ### Added
 * Added default value 100 for argument `n.binom.size` in function `AddNoise`.